move packages around to use as library
