+++ search/testdata/.hiddenDir/dotDir
