+++ search/testdata/symlink
